Create readme.md
Work relating to improvements to either the reliability or sustainability of the egg waste system (DMI). See screenshots from (1) activities relating to improving DMI performance and understanding processing/reliability data and moisture levels and (2) the net-zero report compiled for the same system as part of a separate project with WSP - aimed at scoping the pathway to eliminating direct carbon emissions from this very energy intensive system. These screenshots include a list of the options considered, their energy use and estimates of OPEX. Also attached is an experimental write-up for moisture analysis performed on samples of egg waste at the QC laboratories at Speke, this was to improve our understanding of the system's reliability (1)