Update readme.md
Work relating to improvements to either the reliability or sustainability of the egg waste system (DMI). Includes (1) screenshots of data relating to DMI performance in order to better understand and improve reliability and moisture levels and (2) the net-zero report compiled for the same system as part of a separate project with WSP. This was aimed at scoping the pathway to eliminating direct carbon emissions from this very energy intensive system (~3 million kWh of natural gas use per annum). These screenshots include a list of the options considered, their energy use and estimates of OPEX. Also attached is an experimental write-up for moisture analysis performed on samples of egg waste at the QC laboratories at Speke, this was to improve our understanding of the system's reliability, in line with the data gathering in (1).