Create readme.md
Project proposal, schedule and quotation for contracted work - as part of the project and change control to streamline the site access control system. I arranged and produced the first of these two documents and approached the contractor (Integra) to scope the work required. This project sought to reduce the number of nuisance door alarms on site - each having to be processed individually by our security team. In addition, the work would simplify use of the system for area owners around site and mitigate against registered quality and security risks.