Add files via upload
Screenshots and photos showing the Matterport 3D photo model. From the start of my placement at Speke, I led the introduction of this technology and made rapid progress by co-ordinating and organising training in the technology, scanning the manufacturing area, then developing the resulting model with overlays of useful information (explaining production processes etc.). I worked with colleagues across the engineering and manufacturing departments to produce a fully developed model, with embedded videos and uploaded SOPs/task sheets - a really convenient and comprehensive way to virtually get familiar with a complex labyrinth of machinery and corridors. This was used extensively for the onboarding of ~100 seasonal staff for the 2022 manufacturing campaign.