Create readme.md
Photos and drawings of the Vacuum Transfer/Tablet Packoff Rig. I was the AZ responsible engineer and owner of the delivery of this package under the wider CDC project. In just under 12 months, I was able to progress this machine from a basic design and initial prototype to a production ready version which underwent installation and commissioning, first at the supplier's test facility and finally, at the CDC facility in Gothenburg. I had to ensure that the final product met all AZ client specifications, involving multiple design reviews and supplier conflicts along the way.