Create readme.md
Two major projects during my first year at AstraZeneca. The "Air Tubing Deviation" folder contains images and a 3D model of a protective guard that I designed and 3D printed in response to a critical manufacturing deviation. The guards were to protect a number of vulnerable air hoses. These air hoses were prone to being constricted but served the vital task of taking air from the sterile manufacturing area - to sampling moniters. This process is part of the sterile certification for compliant manufacture of the Zoladex drug. The "Digital Reliability Dashboard" was the other major project. I coded this in Microsoft PowerBI; the dashboard takes data from a complex maintenance system and presents it clearly in such a way to aid decision making. This tool remains in use and helped achieve a 12x decrease in the number of missed maintenance orders across 4 production facilities.