Create readme.md
Images and a 3D model of a protective guard that I designed and 3D printed in response to a critical manufacturing deviation. The guards were to protect a number of vulnerable air hoses. These air hoses were prone to being constricted but served the vital task of taking air from the sterile manufacturing area - to sampling moniters. This process is part of the sterile certification for compliant manufacture of the Zoladex drug. I was able to use modern engineering skills to resolve a critical issue in a live manufacturing area - at very low cost and a fraction of the price of any alternatives.