Create readme.md
I produced this online maintenance dashboard in my first months at AstraZeneca and it completely transformed the approach and performance of the wider maintenance team. The dashboard, coded in Microsoft PowerBI, takes data from a complex maintenance system and presents it clearly in such a way to aid decision making and vastly improve accountability and reliability. Maintenance orders were previously seen in a large excel list that had to be regularly reviewd and scrutinised to see which maintenance orders were on-track and accounted for. This process was very unreliable and prone to human error. The new dashboard presented maintenance orders with the use of graphs and charts that highlighted key issues (lack of an assigned technician, impending due date, overdue status etc) and automatically prioritised work orders. Where previously, maintenance orders were routinely missed due to lack of visibility- they were now planned weeks in advance with a near 100% fullfilment rate. This tool remains in use and helped achieve a 12x decrease in the number of missed maintenance orders across 4 production facilities, in the first year of use.