Add files via upload
Report submitted in advance of the detailed design of a future drug processing plant in Macclesfield.  I estimated and produced an energy target (electric + steam) for the new project, detailing the errors in existing metering and suggestions for improvements in efficiency that could be achieved in the construction of the new facility.